[style] 반응형 기본 스타일 작성 #36
