+++ src/components/Main/ApartInfo/ApartInfo.module.scss
@@ -1,7 +1,7 @@
 @use "@/styles/_colors.scss";
 
 .apartInfoContainer {
-  width: 45%;
+  width: 100%;
   height: 100%;
   display: flex;
   flex-direction: column;
