+++ src/components/Main/ApartMap/ApartMap.module.scss
@@ -1,6 +1,7 @@
 .apartMap {
   border: 0.1rem solid #9e9e9e;
   width: 100%;
+  height: 100%;
   position: relative;
   overflow: hidden;
   .map {
