+++ src/components/Main/InfoContainer.module.scss
@@ -1,0 +1,52 @@
+@use "@/styles/responsive.scss" as rs;
+@use "@/styles/colors.scss";
+
+@mixin testRs {
+  @include rs.respond-to("tablet-large") {
+    position: absolute;
+    bottom: 0;
+    overflow: scroll;
+    z-index: 10;
+    width: 100%;
+    height: 50%;
+    gap: 0;
+    transition: height 0.4s ease-in-out;
+    background-color: #ffffff;
+  }
+}
+
+.info {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  width: 45%;
+  @include testRs();
+  .apartSearch {
+    @include rs.respond-to("tablet-large") {
+      display: none;
+    }
+  }
+  &.closed {
+    @include rs.respond-to("tablet-large") {
+      height: 4rem;
+    }
+  }
+  .slideButton {
+    display: none;
+    @include rs.respond-to("tablet-large") {
+      display: block;
+      background-color: colors.$mainNavy;
+      position: sticky;
+      top: 0;
+      height: 4rem;
+      padding: 1.6rem;
+      border-radius: 2rem;
+      border-bottom-right-radius: 0;
+      border-bottom-left-radius: 0;
+      width: 100%;
+      border: none;
+      cursor: pointer;
+      color: white;
+    }
+  }
+}
