+++ src/pages/ApartInfoPage.module.scss
@@ -1,6 +1,22 @@
-.home {
+.apartInfoPage {
   display: flex;
+  position: relative;
+  overflow: hidden;
   gap: 2.4rem;
   width: 100%;
   height: 100%;
+  // .apartMap {
+  //   width: 100%;
+  //   @include rs.respond-to("tablet-large") {
+  //     display: flex;
+  //     height: 100%;
+  //     flex-direction: column;
+  //   }
+  //   .apartSearch {
+  //     display: none;
+  //     @include rs.respond-to("tablet-large") {
+  //       display: block;
+  //     }
+  //   }
+  // }
 }
