+++ src/pages/News.module.scss
@@ -1,5 +1,15 @@
+@use "@/styles/responsive.scss" as rs;
+
 .news {
   height: 100%;
   display: flex;
   flex-direction: column;
+  @include rs.respond-to("tablet-large") {
+    padding: 2%;
+    box-sizing: border-box;
+  }
+  @include rs.respond-to("mobile-large") {
+    padding: 3%;
+    box-sizing: border-box;
+  }
 }
